+++ spec/models/conversion/task_list_spec.rb
@@ -33,6 +33,7 @@ RSpec.describe Conversion::TaskList do
         :school_completed,
         :conditions_met,
         :share_information,
+        :confirm_date_academy_opened,
         :redact_and_send,
         :update_esfa,
         :receive_grant_payment_certificate
@@ -92,6 +93,7 @@ RSpec.describe Conversion::TaskList do
           {
             identifier: :after_opening,
             tasks: [
+              Conversion::Task::ConfirmDateAcademyOpenedTaskForm,
               Conversion::Task::RedactAndSendTaskForm,
               Conversion::Task::UpdateEsfaTaskForm,
               Conversion::Task::ReceiveGrantPaymentCertificateTaskForm
@@ -118,7 +120,7 @@ RSpec.describe Conversion::TaskList do
       project = create(:conversion_project)
       task_list = described_class.new(project, user)
 
-      expect(task_list.tasks.count).to eql 30
+      expect(task_list.tasks.count).to eql 31
       expect(task_list.tasks.first).to be_a Conversion::Task::HandoverTaskForm
       expect(task_list.tasks.last).to be_a Conversion::Task::ReceiveGrantPaymentCertificateTaskForm
     end